+++ src/main/java/com/example/moviereserve/controller/UserController.java
@@ -8,8 +8,8 @@ import com.example.moviereserve.entity.user.User;
 import com.example.moviereserve.exception.LoginFailureException;
 import com.example.moviereserve.repository.UserRepository;
 import com.example.moviereserve.response.Response;
-import com.example.moviereserve.service.UserSignInService;
-import com.example.moviereserve.service.UserSignUpService;
+import com.example.moviereserve.service.user.UserSignInService;
+import com.example.moviereserve.service.user.UserSignUpService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
